requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# inter-iit-gui/gui.py: 7
Pillow == 7.2.0

# inter-iit-gui/gui.py: 6
numpy == 1.18.5

# inter-iit-gui/gui.py: 5
opencv_python == 4.4.0.44
